produce_tweets/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
pika==1.2.0
PyYAML==6.0
requests==2.27.1
urllib3==1.26.8
